pyproject.toml

[tool.poetry]
name = "challenge-python-v1"
version = "0.1.0"
description = ""
authors = ["Ondrej Ivanko <woshi4ndy@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "*"
pytest = "*"
pillow = "^9.5.0"
djangorestframework = "^3.14.0"
black = "^23.3.0"
pylint = "^2.17.4"
psycopg2 = "^2.9.6"
django-filter = "^23.2"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 79
